Fix: test breaking due to incorrect string appended to the test output
When running the test, the string `-n # TestPackageNamespace` is appended to the dummy test output file `ReplaceTest.md`. Here the `-n ` is not supposed to be here. Also instead of creating only one entry in the file, the replaced text is appended to the dummy file.

On the first run, we will get a failed test with the below output.

```
-n # NewPackageNamespace
```

On next run, this will become:

```
-n # NewPackageNamespace
-n # NewPackageNamespace
```

Here instead of using `-n` we are using `\c` and also instead of appending, we are replacing the content of the file.